requirements.txt
tefas-crawler>=0.5.0
pandas>=2.0.0
